requirements.txt
tensorflow==2.16.1
streamlit
pillow
numpy
